Fixed Mutex and Time Assignment funcitons
Mutex now does not allow for bids on subtasks with mutually exclusive dependencies if such a subtask is mutually exclusive to one in the bundle during phase 1.

Time assignment was generating new initial positions for agents and was returning wrong results for arrival time. This is no longer the case